requirements.txt
discord
python-dotenv
loguru